# Conflicts:
#	client/app/services/auth.service.ts
#	package.json
some minor changes

+++ client/app/admin/admin.component.ts
@@ -3,6 +3,7 @@ import { Component, OnInit } from '@angular/core';
 import { ToastComponent } from '../shared/toast/toast.component';
 import { AuthService } from '../services/auth.service';
 import { UserService } from '../services/user.service';
+import { User } from '../shared/models/user.model';
 
 @Component({
   selector: 'app-admin',
@@ -10,7 +11,7 @@ import { UserService } from '../services/user.service';
 })
 export class AdminComponent implements OnInit {
 
-  users = [];
+  users: User[] = [];
   isLoading = true;
 
   constructor(public auth: AuthService,
@@ -29,12 +30,14 @@ export class AdminComponent implements OnInit {
     );
   }
 
-  deleteUser(user) {
-    this.userService.deleteUser(user).subscribe(
-      data => this.toast.setMessage('user deleted successfully.', 'success'),
-      error => console.log(error),
-      () => this.getUsers()
-    );
+  deleteUser(user: User) {
+    if (window.confirm('Are you sure you want to delete ' + user.username + '?')) {
+      this.userService.deleteUser(user).subscribe(
+        data => this.toast.setMessage('user deleted successfully.', 'success'),
+        error => console.log(error),
+        () => this.getUsers()
+      );
+    }
   }
 
 }
